backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
ebooklib
PyPDF2
requests
beautifulsoup4
python-dotenv
